defininindo as implementacaoes de cada classe e renomeando os packages
