--- src/main/java/br/com/aceleradev/java/centraldeerros/Auth/JWTAuth.java
@@ -1,11 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Auth;
-
-import javax.servlet.*;
-import java.io.IOException;
-
-public class JWTAuth implements Filter {
-  @Override
-  public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-
-  }
-}
