--- src/main/java/br/com/aceleradev/java/centraldeerros/Auth/JWTAuthService.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Auth;
-
-public class JWTAuthService {
-}
