--- src/main/java/br/com/aceleradev/java/centraldeerros/Controller/AccountLoginController.java
@@ -1,22 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Controller;
-
-import br.com.aceleradev.java.centraldeerros.Service.UserAccountService;
-import br.com.aceleradev.java.centraldeerros.Auth.JWTAuthService;
-import br.com.aceleradev.java.centraldeerros.Response.TokenAuthenticationResponse;
-import br.com.aceleradev.java.centraldeerros.Response.CredentialsResponse;
-
-import javax.validation.Valid;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.stereotype.Controller;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.beans.factory.annotation.Autowired;
-
-
-public class AccountLoginController {
-
-  //rota de login
-}
