--- src/main/java/br/com/aceleradev/java/centraldeerros/Controller/UserAccountController.java
@@ -1,35 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Controller;
-
-import br.com.aceleradev.java.centraldeerros.Service.UserAccountService;
-import br.com.aceleradev.java.centraldeerros.Mapper.UserAccountMapper;
-import br.com.aceleradev.java.centraldeerros.Entity.UserAccount;
-import br.com.aceleradev.java.centraldeerros.DTO.UserAccountDTO;
-
-import org.springframework.web.bind.annotation.*;
-import org.springframework.stereotype.Controller;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.HttpStatus;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import lombok.AllArgsConstructor;
-
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-
-import javax.validation.Valid;
-import java.util.Optional;
-
-
-public class UserAccountController {
-
-
-  //criacao do endpoint para salvar conta
-
-  //criacao do endpoiint para buscar uma conta
-
-  //criacao do endpoint para atualizacao da conta
-
-
-
-}
