--- src/main/java/br/com/aceleradev/java/centraldeerros/Repository/ServiceEventRepository.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Repository;
-
-public class ServiceEventRepository {
-}
