--- src/main/java/br/com/aceleradev/java/centraldeerros/Response/CredentialsResponse.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Response;
-
-public class CredentialsResponse {
-}
