--- src/main/java/br/com/aceleradev/java/centraldeerros/Response/TokenAuthenticationResponse.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Response;
-
-public class TokenAuthenticationResponse {
-}
