--- src/main/java/br/com/aceleradev/java/centraldeerros/Rules/BooleanEventRule.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Rules;
-
-public class BooleanEventRule {
-}
