--- src/main/java/br/com/aceleradev/java/centraldeerros/Rules/EnumEventRule.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Rules;
-
-public class EnumEventRule {
-}
