--- src/main/java/br/com/aceleradev/java/centraldeerros/Service/UserAccountService.java
@@ -1,4 +1,0 @@
-package br.com.aceleradev.java.centraldeerros.Service;
-
-public class UserAccountService {
-}
