+++ src/main/java/br/com/aceleradev/java/centraldeerros/auth/JwtAuthorizationFilter.java
@@ -1,0 +1,10 @@
+package br.com.aceleradev.java.centraldeerros.auth;
+
+
+import org.springframework.web.filter.OncePerRequestFilter;
+
+public class JwtAuthorizationFilter extends OncePerRequestFilter {
+
+
+  //impl filter
+}
