+++ src/main/java/br/com/aceleradev/java/centraldeerros/auth/TokenAuthenticationService.java
@@ -1,0 +1,7 @@
+package br.com.aceleradev.java.centraldeerros.auth;
+
+
+public class TokenAuthenticationService {
+
+
+}
