+++ src/main/java/br/com/aceleradev/java/centraldeerros/configuration/ExceptionHandler.java
@@ -1,0 +1,36 @@
+package br.com.aceleradev.java.centraldeerros.configuration;
+
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.stream.Collectors;
+import java.util.Map;
+import java.util.List;
+import java.util.LinkedHashMap;
+import java.time.LocalDateTime;
+
+@ControllerAdvice
+public class ExceptionHandler extends ResponseEntityExceptionHandler {
+
+  @Override
+  protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
+                                                                HttpHeaders headers, HttpStatus status, WebRequest request) {
+
+    Map<String, Object> body = new LinkedHashMap<>();
+    body.put("timestamp", LocalDateTime.now());
+
+    List<String> errors = ex.getBindingResult().getFieldErrors().stream().map(x -> x.getDefaultMessage())
+      .collect(Collectors.toList());
+
+    body.put("errors", errors);
+
+    return new ResponseEntity<>(body, headers, status);
+
+  }
+
+}
