+++ src/main/java/br/com/aceleradev/java/centraldeerros/controller/AccountLoginController.java
@@ -1,0 +1,7 @@
+package br.com.aceleradev.java.centraldeerros.controller;
+
+
+public class AccountLoginController {
+
+  //rota de login
+}
