+++ src/main/java/br/com/aceleradev/java/centraldeerros/controller/ServiceEventController.java
@@ -1,0 +1,18 @@
+package br.com.aceleradev.java.centraldeerros.controller;
+
+import static org.springframework.http.ResponseEntity.status;
+
+public class ServiceEventController {
+
+  //listar eventos
+
+  //buscar evento pelo id
+
+  ////criar novo evento
+
+  //atualizar novo evento
+
+  //excluir evento
+
+
+}
