+++ src/main/java/br/com/aceleradev/java/centraldeerros/repository/ServiceEventRepository.java
@@ -1,0 +1,7 @@
+package br.com.aceleradev.java.centraldeerros.repository;
+
+
+public class ServiceEventRepository {
+
+  //impl findall
+}
