+++ src/main/java/br/com/aceleradev/java/centraldeerros/repository/UserAccountRepository.java
@@ -1,0 +1,15 @@
+package br.com.aceleradev.java.centraldeerros.repository;
+
+
+public class UserAccountRepository {
+
+  //impl findbyid
+
+  //query na tabela useraccount pelo email
+
+  //impl findbyemail
+
+  //impl find by username
+
+
+}
