+++ src/main/java/br/com/aceleradev/java/centraldeerros/response/CredentialsResponse.java
@@ -1,0 +1,9 @@
+package br.com.aceleradev.java.centraldeerros.response;
+
+
+public class CredentialsResponse {
+
+  //adicionar serialversionuuid
+
+  //adicionar validacao notnull de email e password e mensagem de erro
+}
