+++ src/main/java/br/com/aceleradev/java/centraldeerros/response/TokenAuthenticationResponse.java
@@ -1,0 +1,7 @@
+package br.com.aceleradev.java.centraldeerros.response;
+
+
+public class TokenAuthenticationResponse {
+
+  //implementar o uuid e receber o valor no token
+}
