+++ src/main/java/br/com/aceleradev/java/centraldeerros/rules/BooleanEventRule.java
@@ -1,0 +1,8 @@
+package br.com.aceleradev.java.centraldeerros.rules;
+
+
+public class BooleanEventRule {
+
+  //impl jpa specification
+
+}
