+++ src/main/java/br/com/aceleradev/java/centraldeerros/rules/EnumEventRule.java
@@ -1,0 +1,7 @@
+package br.com.aceleradev.java.centraldeerros.rules;
+
+
+public class EnumEventRule {
+
+  //impl jpa specificaton
+}
