+++ src/main/java/br/com/aceleradev/java/centraldeerros/rules/StringEventRule.java
@@ -1,0 +1,8 @@
+package br.com.aceleradev.java.centraldeerros.rules;
+
+
+public class StringEventRule {
+
+  //impl jpa specification
+
+}
