+++ src/main/java/br/com/aceleradev/java/centraldeerros/service/ServiceEventService.java
@@ -1,0 +1,14 @@
+package br.com.aceleradev.java.centraldeerros.service;
+
+public class ServiceEventService {
+
+  //impl findall
+
+  //impl save
+
+  //impl findByid
+
+  //impl update
+
+  //impl delete
+}
