+++ src/main/java/br/com/aceleradev/java/centraldeerros/service/UserAccountService.java
@@ -1,0 +1,16 @@
+package br.com.aceleradev.java.centraldeerros.service;
+
+public class UserAccountService {
+
+  //impl save account
+
+  //impl findbyid
+
+  //impl finbyemail
+
+  //impl hasAccount
+
+  //impl isEmpty
+
+
+}
